+++ packages/legend-graph/src/models/protocols/pure/v1/transformation/pureGraph/to/helpers/V1_MappingBuilderHelper.ts
@@ -60,7 +60,7 @@ import { V1_ObjectInputData } from '../../../../model/packageableElements/store/
 import { V1_FlatDataInputData } from '../../../../model/packageableElements/store/flatData/mapping/V1_FlatDataInputData';
 import type { V1_ClassMapping } from '../../../../model/packageableElements/mapping/V1_ClassMapping';
 import type { V1_MappingInclude } from '../../../../model/packageableElements/mapping/V1_MappingInclude';
-import { V1_resolvePathsInRawLambda } from './V1_RawPathLambdaResolver';
+import { V1_resolvePathsInRawLambda } from './V1_ValueSpecificationPathResolver';
 import { V1_RelationalInputData } from '../../../../model/packageableElements/store/relational/mapping/V1_RelationalInputData';
 import {
   getRelationalInputType,